+++ r-package/R/build_network.R
@@ -83,7 +83,7 @@ build_network <- function(data_path,
     r5r_network <- rJava::.jnew("org.ipea.r5r.R5RCore", data_path, verbose, elevation)
 
     cli::cli_inform(c(
-      i = "Using cached network from from {.path {dat_file}}."
+      i = "Using cached network from {.path {dat_file}}."
     ))
 
   } else {
@@ -124,6 +124,25 @@ build_network <- function(data_path,
       cli::cli_inform(c(
         i = "No public transport data ({.file GTFS}) found in {.path {data_path}}. Graph will be built with the street network only."
       ))
+    }
+
+    errors = java_to_dt(r5r_network$gtfsErrors)
+    errfile = file.path(data_path, "gtfs_errors.csv")
+
+    # always write error file even when empty, so that if you fix errors the error file gets overwritten on rebuild
+    write.csv(errors, errfile)
+
+    if (any(errors$priority == "HIGH")) {
+      cli::cli_abort(c(
+        "x"="High priority GTFS errors found; network build failed.",
+        "i" = "Use {.fn r5r::get_gtfs_errors} to see them, or review {.file {errfile}}."
+
+      ))
+    } else if (nrow(errors) > 0) {
+        cli::cli_inform(c(
+          "!" = "{nrow(errors)} errors found in GTFS.",
+          "i" = "Use {.fn r5r::get_gtfs_errors} to see them, or review {.file {errfile}}."
+        ))
     }
 
     cli::cli_inform(c(