+++ r-package/R/pareto_frontier.R
@@ -132,8 +132,6 @@ pareto_frontier <- function(r5r_network,
 
   # check inputs and set r5r options --------------------------------------
 
-  checkmate::assert_class(r5r_network, "r5r_network")
-  r5r_network <- r5r_network@jcore
   origins <- assign_points_input(origins, "origins")
   destinations <- assign_points_input(destinations, "destinations")
   mode_list <- assign_mode(mode, mode_egress)
@@ -143,6 +141,9 @@ pareto_frontier <- function(r5r_network,
   if (mode_list$transit_mode %like% 'TRANSIT|TRAM|SUBWAY|RAIL|BUS|CABLE_CAR|GONDOLA|FUNICULAR') {
     check_transit_availability_on_date(r5r_network, departure_date = departure$date)
   }
+
+  checkmate::assert_class(r5r_network, "r5r_network")
+  r5r_network <- r5r_network@jcore
 
   max_walk_time <- assign_max_street_time(
     max_walk_time,